test_management_backend/requirements/base.txt
Django>=5.0,<5.1
djangorestframework>=3.14
daphne>=4.1
channels>=4.0
channels-redis>=4.1
psycopg2-binary>=2.9
django-cors-headers>=4.3
django-filter>=24.2
djangorestframework-simplejwt>=5.3
celery>=5.3
redis>=5.0
pydantic>=2.6
python-dotenv>=1.0
Pillow>=10.2
django-extensions>=3.2
whitenoise>=6.6
pyserial>=3.5
pyusb>=1.2
pusher>=3.3
python-json-logger>=2.0
inflection>=0.5
drf-spectacular>=0.27
